search now returns all matching results, programmatic routing broken
