+++ src/components/signUp.vue
@@ -7,7 +7,6 @@
 
 <script>
 import signUpForm from '@/components/signUpForm'
-import desktopHeader from '@/components/header'
 
 export default {
   name: 'signUp',
